services/processing-service/requirements.txt
boto3==1.26.90
PyPDF2==3.0.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0